# spec/openapi.yaml
openapi: 3.1.3
info:
  title: MedVault API
  description: >
    REST API for MedVault — a multi-tenant healthcare platform.
    This contract is the single source of truth for all API endpoints.
  version: 0.1.0
  contact:
    name: MedVault Team
servers:
  - url: /api/v1
    description: API v1
security:
  - bearerAuth: []
...
paths:
...
  /cases/{id}/images/upload-url:
    parameters:
      - $ref: '#/components/parameters/CaseIdParam'

    post:
      tags: [Images]
      operationId: requestUploadURL
      summary: Request pre-signed upload URL
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UploadURLRequest'
      responses:
        '200':
          description: Pre-signed URL generated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadURLResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '403':
          description: Not authorized to upload to this case
        '404':
          $ref: '#/components/responses/NotFound'
  /cases/{id}/images:
    parameters:
      - $ref: '#/components/parameters/CaseIdParam'

    post:
      tags: [Images]
      operationId: confirmUpload
      summary: Confirm image upload and record metadata
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConfirmUploadRequest'
      responses:
        '201':
          description: Image recorded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImageResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFound'

    get:
      tags: [Images]
      operationId: listImages
      summary: List images for case
      responses:
        '200':
          description: List of images
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/ImageResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /images/{id}/download-url:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid

    get:
      tags: [Images]
      operationId: getDownloadURL
      summary: Get pre-signed download URL
      responses:
        '200':
          description: Pre-signed download URL
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DownloadURLResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /images/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid

    delete:
      tags: [Images]
      operationId: deleteImage
      summary: Delete an image (admin only)
      responses:
        '200':
          description: Image deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Only administrators can delete images
        '404':
          $ref: '#/components/responses/NotFound'
  /audit-logs:
    get:
      tags: [Audit]
      operationId: listAuditLogs
      summary: List audit logs (admin only)
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
        - name: action
          in: query
          schema:
            type: string
        - name: user_id
          in: query
          schema:
            type: string
            format: uuid
        - name: resource_type
          in: query
          schema:
            type: string
        - name: resource_id
          in: query
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of audit logs
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/AuditLogResponse'
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Only administrators can view audit logs
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    CaseIdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSizeParam:
      name: page_size
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  responses:
    Unauthorized:
      description: Authentication required
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        data:
          type: object
          nullable: true
        meta:
          $ref: '#/components/schemas/ResponseMeta'
...
    UploadURLRequest:
      type: object
      required: [file_name, content_type, file_size]
      properties:
        file_name:
          type: string
        content_type:
          type: string
          enum: [image/jpeg, image/png, image/dicom]
        file_size:
          type: integer
          format: int64
          minimum: 1
    UploadURLResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            upload_url:
              type: string
              format: uri
            s3_key:
              type: string
            expires_in:
              type: integer
              description: URL lifetime in seconds
        meta:
          $ref: '#/components/schemas/ResponseMeta'
    ConfirmUploadRequest:
      type: object
      required: [s3_key, file_name, content_type]
      properties:
        s3_key:
          type: string
        file_name:
          type: string
        content_type:
          type: string
          enum: [image/jpeg, image/png, image/dicom]
    ImageResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        case_id:
          type: string
          format: uuid
        file_name:
          type: string
        content_type:
          type: string
        uploaded_at:
          type: string
          format: date-time
    DownloadURLResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            download_url:
              type: string
              format: uri
            expires_in:
              type: integer
        meta:
          $ref: '#/components/schemas/ResponseMeta'
    AuditLogResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        tenant_id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        action:
          type: string
        resource_type:
          type: string
        resource_id:
          type: string
          format: uuid
        ip_address:
          type: string
        metadata:
          type: object
          nullable: true
        created_at:
          type: string
          format: date-time
...
    ErrorResponse:
      type: object
      properties:
        data:
          nullable: true
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: array
              items:
                type: object
        meta:
          $ref: '#/components/schemas/ResponseMeta'
    ResponseMeta:
      type: object
      properties:
        request_id:
          type: string
          format: uuid
        timestamp:
          type: string
          format: date-time
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
        page_size:
          type: integer
        total:
          type: integer